requirements.txt
albumentations==1.4.10
lightning==2.3.2
matplotlib==3.9.1
neptune==1.10.4
numpy==1.26.4
opencv-python==4.10.0.84
python-dotenv==1.0.1
rasterio==1.3.10
rich==13.7.1
torch==2.3.1
torchmetrics==1.4.0.post0
torchvision==0.18.1
transformers==4.42.3
tqdm==4.66.4
